+++ src/Parsing-impl.h
@@ -12,7 +12,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.4.2
+  Version: 1.5.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -33,6 +33,7 @@
   1.4.0   K Hoang      07/09/2021 Add support to Portenta H7
   1.4.1   K Hoang      04/10/2021 Change option for PIO `lib_compat_mode` from default `soft` to `strict`. Update Packages Patches
   1.4.2   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
+  1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
  ***************************************************************************************************************************************/
 
 #pragma once
@@ -600,7 +601,7 @@ uint8_t WiFiWebServer::_uploadReadByte(WiFiClient& client)
 
 #else
 
-void WiFiWebServer::_parseArguments(String data)
+void WiFiWebServer::_parseArguments(const String& data)
 {
 
   WS_LOGDEBUG1(F("args: "), data);
@@ -1006,8 +1007,8 @@ bool WiFiWebServer::_parseFormUploadAborted()
 #else
 
 
-bool WiFiWebServer::_parseForm(WiFiClient& client, String boundary, uint32_t len) {
-
+bool WiFiWebServer::_parseForm(WiFiClient& client, const String& boundary, uint32_t len) 
+{
   WS_LOGDEBUG1(F("Parse Form: Boundary: "), boundary);
   WS_LOGDEBUG1(F("Length: "), len);