+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Major Releases v1.5.0](#major-releases-v150)
   * [Releases v1.4.2](#releases-v142)
   * [Releases v1.4.1](#releases-v141)
   * [Major Releases v1.4.0](#major-releases-v140)
@@ -33,6 +34,13 @@
 ---
 
 ## Changelog
+
+### Major Releases v1.5.0
+
+1. Reduce usage of Arduino String with std::string
+2. Optimize library code and examples by using **reference-passing instead of value-passing**.
+3. Update `Packages' Patches`
+4. Add more ESP32/ESP8266 supporting code
 
 ### Releases v1.4.2